bug[hotel, room]:fix hotels model and rooms model association (#56)

+++ src/controllers/hotelController.js
@@ -1,24 +1,42 @@
+/* eslint-disable no-underscore-dangle */
 import model from '../database/models';
-
-require('dotenv').config();
 
 const createHotel = async (req, res) => {
   try {
     const hotel = await model.hotel.create(req.body);
+
     return res.status(201).json({
       hotel,
     });
   } catch (error) {
-    return res.status(500).json({ error: error.message });
+    return res.status(500).json({ message: res.__('Internal server error!') });
   }
 };
 // Getting all hotels
 const getAllHotels = async (req, res) => {
   try {
-    const hotels = await model.hotel.findAll();
+    const hotels = await model.hotel.findAll({
+      include: [
+        {
+          model: model.RoomModel,
+          as: 'rooms',
+          attributes: [
+            'id',
+            'roomType',
+            'description',
+            'roomLabel',
+            'hotelId',
+            'status',
+            'price',
+            'roomImage'
+          ]
+        }
+      ]
+    });
+
     return res.status(200).json({ hotels });
   } catch (error) {
-    return res.status(500).send(error.message);
+    return res.status(500).send({ message: res.__('Internal server error!') });
   }
 };
 // Deleting a particular hotel by id
@@ -28,45 +46,43 @@ const deleteHotel = async (req, res) => {
     const deleted = await model.hotel.destroy({
       where: { id }
     });
+    if (!deleted) return res.status(404).json({ message: res.__('Hotel with the specified ID does not exists') });
     if (deleted) {
-      const deletedRoom = await model.RoomModel.destroy({
-        where: { hotelId : id }
-      })
-      if(deletedRoom)
-        return res.status(200).json({ message: res.__('Hotel deleted successfully.' )});
+      await model.RoomModel.destroy({
+        where: { hotelId: id }
+      });
     }
-    throw new Error('Hotel not found');
+    res.status(200).json({ message: res.__('Hotel deleted successfully.') });
   } catch (error) {
-    return res.status(500).send(error.message);
+    return res.status(500).send({ message: res.__('Internal server error!') });
   }
 };
   // Getting  a particular room by id
 const getHotel = async (req, res) => {
   try {
-    const { id } = req.params;
     const hotel = await model.hotel.findOne({
-      where: { id }
+      where: { id: req.params.id },
+      include: [
+        {
+          model: model.RoomModel,
+          as: 'rooms',
+          attributes: [
+            'id',
+            'roomType',
+            'description',
+            'roomLabel',
+            'hotelId',
+            'status',
+            'price',
+            'roomImage'
+          ]
+        }
+      ]
     });
-    if (hotel) {
-      return res.status(200).json({ hotel });
-    }
-    return res.status(404).send('Hotel with the specified ID does not exists');
+    if (hotel) return res.status(200).json({ hotel });
+    return res.status(404).send({ message: res.__('Hotel with the specified ID does not exists') });
   } catch (error) {
-    return res.status(500).send(error.message);
-  }
-};
-const getHotelRooms = async (req, res) => {
-  try {
-    const { hotelId } = req.params;
-    const rooms = await model.roommodel.findAll({
-      where: { hotelId }
-    });
-    if (rooms) {
-      return res.status(200).json({ rooms });
-    }
-    return res.status(404).send('There is an error !');
-  } catch (error) {
-    return res.status(500).send(error.message);
+    return res.status(500).send({ message: res.__('Internal server error!') });
   }
 };
 const updateHotel = async (req, res) => {
@@ -79,12 +95,12 @@ const updateHotel = async (req, res) => {
       const updatedHotel = await model.hotel.findOne({ where: { id } });
       return res.status(200).json({ hotel: updatedHotel });
     }
-    throw new Error('Hotel not found');
+    return res.status(404).send({ message: res.__('Hotel Not found') });
   } catch (error) {
-    return res.status(500).send(error.message);
+    return res.status(500).send({ message: res.__('Internal server error!') });
   }
 };
   // Exporting functions
 module.exports = {
-  createHotel, getAllHotels, deleteHotel, getHotel, getHotelRooms, updateHotel
+  createHotel, getAllHotels, deleteHotel, getHotel, updateHotel
 };
